RT-Ticket: 5716 RT-Status: resolved RT-Update: correspond
Duplicate code change backed out of 3.2 and checked into 3.3

git-svn-id: svn+ssh://svn.bestpractical.com/svn/bps-public/rt/branches/3.3-TESTING@1345 e417ac7c-1bcc-0310-8ffa-8f5827389a85

+++ lib/RT/I18N.pm
@@ -360,7 +360,6 @@ sub _FindOrGuessCharset {
 
 # }}}
 
-
 # {{{ _GuessCharset
 
 =head2 _GuessCharset STRING
@@ -426,6 +425,7 @@ sub SetMIMEHeadToEncoding {
     return if $charset eq $enc and $preserve_words;
 
     foreach my $tag ( $head->tags ) {
+        next unless $tag; # seen in wild: headers with no name
         my @values = $head->get_all($tag);
         $head->delete($tag);
         foreach my $value (@values) {
